+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Maps unknown `state` values to `state` options for `ARG` country.
+
 ## [3.16.9] - 2021-06-10
 
 ## Fixed